requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
PyJWT==2.4.0
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
